+++ IntrospectTests/UIKitTests.swift
@@ -217,7 +217,7 @@ private struct TextFieldTestView: View {
     }
 }
 
-@available(iOS 14.0, macCatalyst 14.0, macOS 11.0, *)
+@available(iOS 14.0, macCatalyst 14.0, macOS 11.0, tvOS 13.0, *)
 @available(tvOS, unavailable, message: "TextEditor is not available in tvOS.")
 private struct TextEditorTestView: View {
     let spy: () -> Void
@@ -297,6 +297,20 @@ private struct SegmentedControlTestView: View {
         }
         .pickerStyle(SegmentedPickerStyle())
         .introspectSegmentedControl { segmentedControl in
+            self.spy()
+        }
+    }
+}
+
+@available(iOS 14.0, tvOS 13.0, macOS 11.0, *)
+@available(tvOS, unavailable)
+private struct ColorWellTestView: View {
+    @State private var color = Color.black
+    let spy: () -> Void
+    
+    var body: some View {
+        ColorPicker("Picker", selection: $color)
+        .introspectColorWell { colorWell in
             self.spy()
         }
     }
@@ -512,12 +526,24 @@ class UIKitTests: XCTestCase {
         wait(for: [expectation], timeout: TestUtils.Constants.timeout)
     }
     
-    @available(iOS 14.0, macCatalyst 14.0, macOS 15.0, *)
+    @available(iOS 14.0, macCatalyst 14.0, macOS 11.0, *)
     @available(tvOS, unavailable, message: "TextEditor is not available in tvOS.")
     func testTextEditor() {
 
         let expectation = XCTestExpectation()
         let view = TextEditorTestView(spy: {
+            expectation.fulfill()
+        })
+        TestUtils.present(view: view)
+        wait(for: [expectation], timeout: TestUtils.Constants.timeout)
+    }
+    
+    @available(iOS 14.0, macCatalyst 14.0, macOS 11.0, *)
+    @available(tvOS, unavailable, message: "ColorPicker is not available in tvOS.")
+    func testColorPicker() {
+
+        let expectation = XCTestExpectation()
+        let view = ColorWellTestView(spy: {
             expectation.fulfill()
         })
         TestUtils.present(view: view)